backend/requirements.txt
Flask==2.0.1
Flask-Cors==3.0.10
Flask-Uploads==0.2.1
PyPDF2==1.26.0
Pillow==8.2.0
python-magic==0.4.27
requests==2.25.1